python/fastapi_demo/requirements.txt
fastapi==0.85.0
ujson
uvicorn[standard]
requests
websocket-client
